# api/openapi.yaml
openapi: 3.1.0
info:
  title: Redy Agents API
  version: 0.2.0
  description: Java service with an in-memory session store and demo model.
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        '200':
          description: Healthy
  /v1/agents/sessions:
    post:
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent]
              properties:
                agent:
                  $ref: '#/components/schemas/AgentConfig'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /v1/agents/sessions/{sessionID}:
    parameters:
      - $ref: '#/components/parameters/sessionID'
    get:
      operationId: getSession
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/agents/sessions/{sessionID}/turns:
    parameters:
      - $ref: '#/components/parameters/sessionID'
    post:
      operationId: startTurn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  minLength: 1
      responses:
        '202':
          description: Turn started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /v1/agents/sessions/{sessionID}/turns/{turnID}:
    parameters:
      - $ref: '#/components/parameters/sessionID'
      - $ref: '#/components/parameters/turnID'
    get:
      operationId: getTurn
      responses:
        '200':
          description: Turn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/agents/sessions/{sessionID}/turns/{turnID}/cancel:
    parameters:
      - $ref: '#/components/parameters/sessionID'
      - $ref: '#/components/parameters/turnID'
    post:
      operationId: cancelTurn
      responses:
        '200':
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/agents/sessions/{sessionID}/events:
    parameters:
      - $ref: '#/components/parameters/sessionID'
      - name: after
        in: query
        description: Return events with a sequence greater than this cursor.
        schema:
          type: integer
          minimum: 0
      - name: Last-Event-ID
        in: header
        description: SSE reconnect cursor, used when after is omitted.
        schema:
          type: string
    get:
      operationId: getEvents
      responses:
        '200':
          description: JSON snapshot or SSE stream, selected by Accept header.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    sessionID:
      name: sessionID
      in: path
      required: true
      schema:
        type: string
    turnID:
      name: turnID
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Session already has an active turn, or turn cannot be cancelled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    AgentConfig:
      type: object
      required: [name, model]
      properties:
        name:
          type: string
        model:
          type: string
          description: Only demo is supported by the bundled model adapter.
        instructions:
          type: string
    Session:
      type: object
      required: [id, agent, status, created_at, updated_at]
      properties:
        id:
          type: string
        agent:
          $ref: '#/components/schemas/AgentConfig'
        status:
          type: string
          enum: [idle, running]
        active_turn_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Turn:
      type: object
      required: [id, session_id, input, status, created_at]
      properties:
        id:
          type: string
        session_id:
          type: string
        input:
          type: string
        status:
          type: string
          enum: [running, cancelling, completed, failed, cancelled]
        output:
          type: string
        error:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Event:
      type: object
      required: [sequence, type, session_id, created_at]
      properties:
        sequence:
          type: integer
        type:
          type: string
        session_id:
          type: string
        turn_id:
          type: string
        created_at:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string